Allow for Model Switching
This adds a feature for the user to switch the model used. If
there is only one model, the selector disappears from the UI.

If there are are more than one, a selector appears with the name
of the model, and upon being selected if a description has been
given that will appear on the screen.

Resolves #2