Input Should be In Meters
The UI labels say meters but the scaling of the layers was based
on kilometers, this is now rectified, mostly.

We're still using a sort of guestimated constant here based on
some techniques I've found on the intarwebs for constantly scaling
a radius to a zoom level.  The issue is most of them seem to
indicate that the constant should be the initial zoom level, but
in my experiments this doesn't work.  This may because they're
using a different drawing technique than I am, dunno.

The constant I have found seems to be roughly correct.  If we find
we want more accuracy as far as the layer boundries, I can
investigate this further.

Resolves #9